+++ kratos/middleware/otel/config_test.go
@@ -1,4 +1,4 @@
-package tracing
+package otel
 
 import (
 	"testing"
