+++ kratos/middleware/otel/doc.go
@@ -1,0 +1,3 @@
+// Package otel provides OpenTelemetry tracing middleware for Kratos clients and
+// servers.
+package otel
