+++ kratos/middleware/otel/internal/buf.yaml
