+++ kratos/middleware/otel/metadata.go
@@ -1,4 +1,4 @@
-package tracing
+package otel
 
 import (
 	"context"
