+++ kratos/middleware/otel/middleware.go
@@ -1,9 +1,9 @@
-package tracing
+package otel
 
 import (
 	"context"
 
-	"github.com/go-fries/fries/kratos/middleware/otel/tracing/v3/internal/semconv"
+	"github.com/go-fries/fries/kratos/middleware/otel/v3/internal/semconv"
 	"github.com/go-kratos/kratos/v2/middleware"
 	"github.com/go-kratos/kratos/v2/transport"
 	"go.opentelemetry.io/otel/trace"
