+++ kratos/middleware/otel/middleware_test.go
@@ -1,4 +1,4 @@
-package tracing
+package otel
 
 import (
 	"context"
