+++ kratos/middleware/otel/statshandler.go
@@ -1,9 +1,9 @@
-package tracing
+package otel
 
 import (
 	"context"
 
-	"github.com/go-fries/fries/kratos/middleware/otel/tracing/v3/internal/semconv"
+	"github.com/go-fries/fries/kratos/middleware/otel/v3/internal/semconv"
 	"go.opentelemetry.io/otel/trace"
 	"google.golang.org/grpc/peer"
 	"google.golang.org/grpc/stats"
