+++ kratos/middleware/otel/statshandler_test.go
@@ -1,4 +1,4 @@
-package tracing
+package otel
 
 import (
 	"context"
