+++ kratos/middleware/otel/tracer.go
@@ -1,17 +1,17 @@
-package tracing
+package otel
 
 import (
 	"context"
 	"fmt"
 
-	"github.com/go-fries/fries/kratos/middleware/otel/tracing/v3/internal/semconv"
+	"github.com/go-fries/fries/kratos/middleware/otel/v3/internal/semconv"
 	"github.com/go-kratos/kratos/v2/errors"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
 )
 
-const scopeName = "github.com/go-fries/fries/kratos/middleware/otel/tracing/v3"
+const scopeName = "github.com/go-fries/fries/kratos/middleware/otel/v3"
 
 type tracer struct {
 	tracer     trace.Tracer
