--- kratos/middleware/otel/tracing/doc.go
@@ -1,3 +1,0 @@
-// Package tracing provides OpenTelemetry tracing middleware for Kratos clients
-// and servers.
-package tracing
