+++ kratos/middleware/otel/version.go
@@ -1,4 +1,4 @@
-package tracing
+package otel
 
 // Version returns the current release version of this instrumentation.
 func Version() string {
