+++ kratos/middleware/otel/version_test.go
@@ -1,10 +1,10 @@
-package tracing_test
+package otel_test
 
 import (
 	"regexp"
 	"testing"
 
-	"github.com/go-fries/fries/kratos/middleware/otel/tracing/v3"
+	"github.com/go-fries/fries/kratos/middleware/otel/v3"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -15,6 +15,6 @@ var versionRegex = regexp.MustCompile(`^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)
 	`(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$`)
 
 func TestVersionSemver(t *testing.T) {
-	v := tracing.Version()
+	v := otel.Version()
 	assert.NotNil(t, versionRegex.FindStringSubmatch(v), "version is not semver: %s", v)
 }
